requirements-dev.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
astroid==3.3.10; python_full_version >= '3.9.0'
authlib==1.6.0; python_version >= '3.9'
bandit==1.8.3; python_version >= '3.9'
black==25.1.0; python_version >= '3.9'
certifi==2025.4.26; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
cfgv==3.4.0; python_version >= '3.8'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
coverage[toml]==7.8.1; python_version >= '3.9'
cryptography==45.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
dill==0.4.0; python_version >= '3.8'
distlib==0.3.9
dparse==0.6.4; python_version >= '3.7'
filelock==3.16.1; python_version >= '3.8'
flake8==6.1.0
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
identify==2.6.10; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
iniconfig==2.1.0; python_version >= '3.8'
isort==6.0.1; python_full_version >= '3.9.0'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.5.0; python_version >= '3.9'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==4.0.0; python_version >= '3.9'
mccabe==0.7.0; python_version >= '3.6'
mdurl==0.1.2; python_version >= '3.7'
mypy==1.11.2
mypy-extensions==1.1.0; python_version >= '3.8'
nltk==3.9.1; python_version >= '3.8'
nodeenv==1.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==25.0; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pbr==6.1.1; python_version >= '2.6'
platformdirs==4.3.8; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
pre-commit==4.2.0; python_version >= '3.9'
psutil==7.0.0; python_version >= '3.6'
pycodestyle==2.13.0; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pyflakes==3.3.2; python_version >= '3.9'
pygments==2.19.1; python_version >= '3.8'
pylint==3.3.7; python_full_version >= '3.9.0'
pytest==8.3.5; python_version >= '3.8'
pytest-cov==6.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rich==14.0.0; python_full_version >= '3.8.0'
ruamel-yaml==0.18.10; python_version >= '3.7'
ruamel.yaml.clib==0.2.12; python_version >= '3.9'
safety==3.5.1; python_version >= '3.8'
safety-schemas==0.0.14; python_version >= '3.7'
setuptools==80.8.0; python_version >= '3.9'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
stevedore==5.4.1; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
tomlkit==0.13.2; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typer==0.15.4; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
urllib3==2.4.0; python_version >= '3.9'
virtualenv==20.31.2; python_version >= '3.8'
appdirs==1.4.4
caldav==1.4.0
icalendar==6.3.1; python_version >= '3.8'
lxml==5.4.0; python_version >= '3.6'
pillow==11.2.1; python_version >= '3.9'
plyer==2.1.0
pyobjc-core==11.0; python_version >= '3.8'
pyobjc-framework-cocoa==11.0; python_version >= '3.9'
pyobjc-framework-quartz==11.0; python_version >= '3.9'
pyobjus==1.2.3
pyqt6==6.9.0; python_version >= '3.9'
pyqt6-qt6==6.9.0
pyqt6-sip==13.10.0; python_version >= '3.9'
pystray==0.19.5
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
recurring-ical-events==3.7.0; python_version >= '3.8'
simpleaudio==1.0.4
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
tzdata==2025.2; python_version >= '2'
vobject==0.9.9
x-wr-timezone==2.0.1; python_version >= '3.9'
types-PyYAML==6.0.12.20250516
types-pytz==2025.2.0.20250516
types-psutil==7.0.0.20250516

# Build tools
